# Cache Exif data