Fixes #6291